THIS IS MERELY FOR LOWPOP
Adds in a Great Furnace to the smeltery of the Blacksmith. This is ONLY to combat the fact that with current pop nobody actively goes out of their way to get a riddle, so people are less likely to want to play blacksmith. I'll probably change this back if pop becomes sizable enough.